render_tree: Fix indentation when rendering from node other than "true" root
If we're rendering a subtree, we want to indent from the root of the
subtree, not the root of the whole tree (since we can't see the rest of the
tree). We could simply remove the subtree's parent and effectively cut it
off, but that would break the paths of the nodes within it. Instead we'll
just remember which node we're rendering from, and indent as if that was
the root.